Fixed error on search box searching for null
we were loading from loaded list which ended up being null when the search box was empty so we were throwing an error. That's fixed, but now we just have a bug that when we search for a todo for the first time we always have an empty list. Which we can fix later !